Reduntant method moved to abstract (super) class and "Period" value is removed from the HourlyPrediction class.
