+++ io.openems.edge.predictor.api/src/io/openems/edge/predictor/api/HourlyPrediction.java
@@ -8,14 +8,12 @@ public class HourlyPrediction {
 	//ArrayList<Integer> values = new ArrayList<Integer>(24);
 	public Integer[] values = new Integer[24];
 	public LocalDateTime start;
-	// String period doesn't makes sense
-	public String period = new String("1 hr");
+
 	
-	public HourlyPrediction(Integer[] values, LocalDateTime start, String period) {
+	public HourlyPrediction(Integer[] values, LocalDateTime start) {
 		super();
 		this.values = values;
-		this.start = start;
-		this.period = period;
+		this.start = start;		
 	}
 
 }
