+++ io.openems.edge.predictor.persistant.model/src/io/openems/edge/predictor/persistant/consumption/ConsumptionPersistantModelPredictor.java
@@ -1,8 +1,6 @@
 package io.openems.edge.predictor.persistant.consumption;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Map.Entry;
 import java.util.TreeMap;
 
 import org.osgi.service.component.ComponentContext;
@@ -21,8 +19,6 @@ import io.openems.edge.common.component.ComponentManager;
 import io.openems.edge.common.component.OpenemsComponent;
 import io.openems.edge.common.event.EdgeEventConstants;
 import io.openems.edge.predictor.api.ConsumptionHourlyPredictor;
-import io.openems.edge.predictor.api.HourlyPrediction;
-import io.openems.edge.predictor.api.ProductionHourlyPredictor;
 import io.openems.edge.predictor.persistant.model.AbstractPersistentModelPredictor;
 
 @Designate(ocd = Config.class, factory = true)
@@ -41,23 +37,6 @@ public class ConsumptionPersistantModelPredictor extends AbstractPersistentModel
 
 	public ConsumptionPersistantModelPredictor() {
 		super("_sum/ConsumptionActiveEnergy");
-	}
-
-	@Override
-	public HourlyPrediction get24hPrediction() {
-
-		hourlyEnergyData = super.hourlyEnergyData;
-
-		ArrayList<Long> valueList = new ArrayList<Long>();
-
-		if (!hourlyEnergyData.isEmpty()) {
-			for (Entry<LocalDateTime, Long> entry : hourlyEnergyData.entrySet()) {
-				valueList.add(entry.getValue());
-			}
-		}
-		Integer[] val = valueList.toArray(new Integer[valueList.size()]);
-		HourlyPrediction hourlyPrediction = new HourlyPrediction(val, hourlyEnergyData.firstKey(), "1 hr");
-		return hourlyPrediction;
 	}
 
 	@Activate
